Change naming back to record
