+++ src/slot.rs
@@ -18,7 +18,7 @@ impl<F: FnMut(Signal)> Slot for F {
 
 /// The signal used for communicating state changes.
 ///
-/// For example, if the timeline can no longer redo any actions,
+/// For example, if the history tree can no longer redo any actions,
 /// it sends a `Redo(false)` signal to tell the user.
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
